requirements.txt
pillow
fitz
PyMuPDF
faiss
openai==0.28.0
